morpheus/requirements.txt
uvicorn
fastapi
sentry-sdk
arq
foxglove-web
aiofiles
ujson
sqlalchemy
psycopg2
markupsafe
chevron
phonenumbers
ua_parser
